+++ app/controllers/elements_controller.rb
@@ -1,4 +1,12 @@
 class ElementDecorator < EntryDecorator
+  class << self
+    def column_names
+      [:sid, :competition_name, :date, :season,
+       :ranking, :skater_name, :nation,
+       :number, :name, :credit, :info, :base_value, :goe, :judges, :value,
+      ]
+    end
+  end
   def sid
     h.link_to_score(nil, model.score)
   end
@@ -40,25 +48,6 @@ class ElementsController < ApplicationController
         col.where(arel)
       }
     }.merge(score_filters)
-=begin
-    @_filteres ||= IndexFilters.new.tap do |f|
-      f.filters = {
-        name: {
-          operator: (params[:perfect_match]) ? :eq : :like,
-          input: :text_field, model: Element,
-        },
-        perfect_match: { operator: nil, input: :checkbox, value: 'PERFECT_MATCH'},
-        goe: { operator: :compare, input: :text_field, model: Element},
-            
-      }.merge score_filters.filters
-    end
-=end
-  end
-  def display_keys
-    [:sid, :competition_name, :date, :season,
-     :ranking, :skater_name, :nation,
-     :number, :name, :credit, :info, :base_value, :goe, :judges, :value,
-    ]
   end
   def collection
     filter(Element.includes(:score, score: [:competition]))    #.filter(filters.create_arel_tables(params)).select("scores.*, competitions.season, elements.*")