+++ app/decorators/category_result_decorator.rb
@@ -6,7 +6,7 @@ class CategoryResultDecorator < EntryDecorator
     as_score(model.scores.first.try(:tss))
   end
   def free_tss
-    as_score(model.scores.second.try(:tss))
+    as_score(model.scores.first.try(:tss))
   end
   self.display_as(:ranking, [:short_ranking, :free_ranking])
   self.display_as(:score, [:points])