+++ app/decorators/skater_decorator.rb
@@ -1,4 +1,9 @@
 class SkaterDecorator < EntryDecorator
+  class << self
+    def column_names
+      [ :name, :nation, :category, :isu_number]
+    end
+  end
   def name
     h.link_to_skater(model)
   end